Final Commit - Componentized more - Fixed moon/sun icon issue - Fixed gradient bg import - Add some mobile view css
